requirements/dev.txt
-r base.txt
ipython
ipdb
django-nose
django-extensions
Werkzeug
django-debug-toolbar
coverage
factory-boy
